修改首页界面，点击可进行跳转

+++ src/pages/complain.vue
@@ -29,13 +29,24 @@
         </div>
       </template>
     </el-table-column>
-    <el-table-column align="center" prop="postId" width="100" label="举报帖子Id" />
+    <el-table-column align="center" prop="postId" width="120" label="被举报帖子Id" />
     <el-table-column width="400" align="center" prop="content" label="举报内容" />
     <el-table-column prop="tag" width="100" label="举报标签" />
     <el-table-column fixed="right" label="操作">
       <template #default="scope">
-        <el-button link type="primary" size="small" @click="handleClick(scope.row.postId)"
-          >删除该帖子</el-button
+        <el-button
+          link
+          type="primary"
+          size="small"
+          @click="handleClick(scope.row.postId, scope.row.reported_id)"
+          >确认违规</el-button
+        >
+        <el-button
+          link
+          type="primary"
+          size="small"
+          @click="handleDelete(scope.row.postId)"
+          >没有违规</el-button
         >
       </template>
     </el-table-column>
@@ -44,24 +55,71 @@
 
 <script setup>
 import { ref, onMounted, onUnmounted } from 'vue'
-import { getComplain, deleteComplain } from '../api/complain'
-
+import { getComplain, deleteComplain, deletePost, addRepotedTotal } from '../api/complain'
+import { ElMessageBox, ElMessage } from 'element-plus'
 const tableData = ref([])
-const handleClick = (postId) => {
-  const params = { postId }
-  deleteComplain(params)
-    .then((res) => {
-      getComplain()
+const handleDelete = (postId) => {
+  ElMessageBox.confirm('确定该违规帖子无违规并且删除记录？', '提示', {
+    confirmButtonText: '确定',
+    cancelButtonText: '取消',
+    type: 'warning',
+  })
+    .then(() => {
+      const params = { postId }
+      // 删除举报表中的帖子
+      deletePost(params)
+        .then((res) => {
+          getcomplain()
+          ElMessage.success('删除成功')
+        })
+        .catch((error) => {
+          console.error('删除失败', error)
+        })
     })
     .catch((error) => {
       console.error('删除失败', error)
     })
 }
 
+const handleClick = (postId, reported_id) => {
+  ElMessageBox.confirm('删除该违规帖子并且记录此人的违规次数？', '提示', {
+    confirmButtonText: '确定',
+    cancelButtonText: '取消',
+    type: 'warning',
+  })
+    .then(() => {
+      const params = { postId }
+      //删除前端的帖子
+      deleteComplain(params)
+        .then((res) => {
+          //删除举报表中的帖子
+          deletePost(params).then((res) => {
+            //举报人加一
+            const param = { reported_id }
+            addRepotedTotal(param)
+              .then((res) => {
+                getcomplain()
+              })
+              .catch((error) => {
+                console.error('删除失败', error)
+              })
+          })
+
+          ElMessage.success('删除成功')
+        })
+        .catch((error) => {
+          console.error('删除失败', error)
+        })
+    })
+    .catch(() => {
+      // 用户点击取消按钮时的操作，可以不做任何处理
+    })
+}
+
 const getcomplain = () => {
   getComplain()
     .then((res) => {
-      tableData.value.push(...res.data)
+      tableData.value = res.data
     })
     .catch((error) => {
       console.error('Error fetching post:', error)
